Properly handle errors from the fuse device.
At the moment, we are indiscriminately returning a SessionFailure("epoll error")
if any of the registered file descriptors would return an error upon read. This
means that the more detailed error handling code below can never be reached.

Instead, attempt to read from fuse device when the device is either
readable *or* has an error state, and then handle the error as we always
intended.

Fixes: #197.